Refactor storage safety heuristics
