+++ src-tauri/src/file_cleaner/safety/mod.rs
@@ -1,0 +1,6 @@
+mod context;
+mod policy;
+mod risk;
+
+pub(crate) use policy::{calculate_safety_score, policy_for_category};
+pub(crate) use risk::{assess_path_risk, RiskLevel};
